backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-rest-swagger==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
Markdown==3.8
MarkupSafe==3.0.2
mysqlclient==2.2.7
openapi-codec==1.3.2
packaging==25.0
pluggy==1.6.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.20.1
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
